+++ src/main/java/org/project/repository/BookRepository.java
@@ -24,5 +24,5 @@ public interface BookRepository extends JpaRepository<Book, Long> {
             "OR CONCAT('', b.ISBN) LIKE %:str%")
     public Iterable<Book> findByAllColumns(@Param("str") String str);
 
-    public List<Book> findByISBNNot(int isbn);
+    public List<Book> findByISBNNot(long isbn);
 }
